fix: FreeRunningView에 포함되어있던 viewModel 변경 로직 FreeRunningViewModel로 이전
